requirements.txt
contourpy>=1.2.0
cycler==0.12.1
fonttools==4.45.1
imageio==2.33.0
importlib-resources==6.1.1
kiwisolver==1.4.5
lazy_loader==0.3
matplotlib==3.8.2
networkx==3.2.1
numpy==1.26.2
opencv-python==4.8.1.78
packaging==23.2
Pillow==10.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
scikit-image==0.22.0
scipy==1.11.4
six==1.16.0
tifffile==2023.9.26
tqdm==4.66.1
zipp==3.17.0
